feat 당첨 통계 로직 및 출력 뷰 구현

+++ src/main/java/lotto/domain/LottoMachine.java
@@ -12,7 +12,7 @@ public class LottoMachine {
 
     public Lottos buy(Money money, LottoGenerator lottoGenerator) {
         checkMoney(money);
-        int lottoCount = money.divideBy(LOTTO_PRICE);
+        int lottoCount = (int) money.divideBy(LOTTO_PRICE);
 
         List<Lotto> lottoList = IntStream.range(0, lottoCount)
                 .mapToObj((num) -> lottoGenerator.generate())
